be more Glimmer-friendly
page-item was setting `current` and expecting that to bubble up to
`pagination-pager`. In Glimmer, that's a no-no; data flows down,
actions up.

Instead, we simply trigger the change event and let the parent
`pagination-pager` component respond and change the data.

This also changes `page-item`'s action from `setSelected`
to `select` to reflect that it's not *setting* anything directly.